+++ src/Persistence/EntityFramework/Migrations/EntityDataContextModelSnapshot.cs
@@ -808,7 +808,7 @@ namespace MUnique.OpenMU.Persistence.EntityFramework.Migrations
 
                     b.Property<string>("Name");
 
-                    b.Property<byte>("Number");
+                    b.Property<short>("Number");
 
                     b.Property<Guid?>("SkillId");
